Move ContentItems into own module folder
