+++ src/components/StageEditor/sections/ContentItems/index.js
@@ -1,0 +1,1 @@
+export { default } from './ContentItems';
